refactored styles
