+++ packages-kintone-customize/app-229-contracts/src/view/indexShow/renderSchedule/sections/results/resultTable/monthRow/MonthRowActual.tsx
@@ -1,7 +1,8 @@
-import { TableCell, TableRow } from '@mui/material';
+import { TableCell } from '@mui/material';
 import { GroupedContracts } from '../../../../hooks/groupContracts';
 import { projTypesToShow } from '../../../../config';
-import { ContractsCell } from '../tableComponents/ContractsCell';
+import { ContractsCell } from './common/ContractsCell';
+import { Fragment } from 'react';
 
 
 
@@ -19,17 +20,7 @@ export const MonthRowActual = ({
   } = contractsData ?? {};
 
   return (
-    <TableRow 
-      sx={{
-
-        '& .MuiTableCell-root:not(:first-of-type)': {
-          fontWeight: 'bold',
-          fontSize: 24,
-          color: 'red',
-          textAlign: 'right',
-        },
-      }}
-    >
+    <Fragment>
       <TableCell>
         実績値
       </TableCell>
@@ -37,27 +28,32 @@ export const MonthRowActual = ({
         id,
       }) => {
         return (
-          <ContractsCell key={id} values={contractsByType?.[id]} />
+          <ContractsCell 
+            key={id} 
+            values={contractsByType?.[id]}
+            color='red'
+          />
         );
       })}
 
 
       {/* その他 */}
-      <ContractsCell values={contractsByType?.['その他']} />
+      <ContractsCell 
+        values={contractsByType?.['その他']}
+        color='red'
+      />
  
       {/* 合計 */}
-      <ContractsCell values={{
-        data: contracts,
-        totalAmtExclTax,
-      }}
+      <ContractsCell
+        values={{
+          data: contracts,
+          totalAmtExclTax,
+        }}
+        color='red'
+
       />
 
-      <TableCell>
-        -
-      </TableCell>
-
-
-    </TableRow>
+    </Fragment>
 
   );
