+++ packages-kintone-customize/app-229-contracts/src/view/indexShow/renderSchedule/sections/results/resultTable/monthRow/common/StyledCell.tsx
@@ -1,0 +1,8 @@
+import { TableCell, styled } from '@mui/material';
+
+
+export const StyledCell = styled(TableCell)`
+    font-weight: bold;
+    font-size: 20px;
+    text-align: right;
+`;
